Moving DelegateAuthInterceptor to a higher number module to be able to use it from 980-commons
