+++ 260-delegate/src/main/java/io/harness/delegate/service/core/client/BUILD.bazel
@@ -18,6 +18,7 @@ java_library(
         "//980-commons/src/main/java/io/harness/exception:module",
         "//980-commons/src/main/java/io/harness/network:module",
         "//980-commons/src/main/java/io/harness/security:module",
+        "//980-commons/src/main/java/io/harness/security/delegate:module",
         "//980-commons/src/main/java/io/harness/serializer/kryo:module",
         "//980-commons/src/main/java/io/harness/version:module",
         "//999-annotations/src/main/java/io/harness/annotations/dev:module",
