+++ 260-delegate/src/main/java/io/harness/eventpublisherclient/EventPublisherClientFactory.java
@@ -12,7 +12,6 @@ import io.harness.annotations.dev.OwnedBy;
 import io.harness.delegate.configuration.DelegateConfiguration;
 import io.harness.exception.KeyManagerBuilderException;
 import io.harness.exception.SslContextBuilderException;
-import io.harness.managerclient.DelegateAuthInterceptor;
 import io.harness.network.FibonacciBackOff;
 import io.harness.network.Http;
 import io.harness.network.NoopHostnameVerifier;
@@ -20,6 +19,7 @@ import io.harness.security.TokenGenerator;
 import io.harness.security.X509KeyManagerBuilder;
 import io.harness.security.X509SslContextBuilder;
 import io.harness.security.X509TrustManagerBuilder;
+import io.harness.security.delegate.DelegateAuthInterceptor;
 import io.harness.serializer.kryo.DelegateKryoConverterFactory;
 import io.harness.version.VersionInfoManager;
