+++ 260-delegate/src/main/java/io/harness/verificationclient/BUILD.bazel
@@ -15,6 +15,7 @@ java_library(
         "//957-cg-beans:module",
         "//970-ng-commons:module",
         "//980-commons:module",
+        "//980-commons/src/main/java/io/harness/security/delegate:module",
         "//999-annotations/src/main/java/io/harness/annotations/dev:module",
         "//srm-service/modules/cv-nextgen-service/contracts/delegate:module",
         "@maven//:com_google_inject_guice",
