+++ 420-delegate-agent/src/main/java/io/harness/managerclient/BUILD.bazel
@@ -26,6 +26,7 @@ java_library(
         "//980-commons/src/main/java/io/harness/network:module",
         "//980-commons/src/main/java/io/harness/rest:module",
         "//980-commons/src/main/java/io/harness/security:module",
+        "//980-commons/src/main/java/io/harness/security/delegate:module",
         "//980-commons/src/main/java/io/harness/serializer/kryo:module",
         "//980-commons/src/main/java/io/harness/version:module",
         "//999-annotations/src/main/java/io/harness/annotations/dev:module",
